RangeFixer : add a caution on the feature when a range is out of bound
SequenceChangedListener : update the interaction when adding the caution

UniprotProteinUpdater added a new method to manage out of bound ranges

RangeChecker : it is not the rangeChecker which will check out of bound ranges

ProteinServiceImpl : if a range is out of bound, will not update the sequence of the protein. The same for protein transcripts.